chore: make libSGM installable in editable mode
When installing in editable mode (i.e. with `-e` option), setuptools
tries to copy builded `.so` file of `sgm_wrapper` into source directory.
But, as in `Extension` declaration in the `setup.py`, `sgm_wrapper`
module is declared to be part of the `libSGM` package, setuptools
consider source directory to be `libSGM` which does not exist. Then
installation in editable mode fails at this step.

In this commit, repository layout is changed to `src-layout` with
`sgm_wrapper` sources put into a `libSGM` package.

For reference, see [setuptools
documentation](https://setuptools.pypa.io/en/latest/userguide/package_discovery.html#package-discovery-and-namespace-packages)

Refs: #68